+++ lib/src/internal/copy_on_write_map.dart
@@ -17,11 +17,7 @@ class CopyOnWriteMap<K, V> implements Map<K, V> {
   V operator [](Object key) => _map[key];
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  Map<K2, V2> cast<K2, V2>() {
-    throw new UnimplementedError('cast');
-  }
+  Map<K2, V2> cast<K2, V2>() => new CopyOnWriteMap<K2, V2>(_map.cast<K2, V2>());
 
   @override
   bool containsKey(Object key) => _map.containsKey(key);
@@ -30,13 +26,7 @@ class CopyOnWriteMap<K, V> implements Map<K, V> {
   bool containsValue(Object value) => _map.containsValue(value);
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_getter
-  Iterable<Null> get entries {
-    // Change Iterable<Null> to Iterable<MapEntry<K, V>> when
-    // the MapEntry class has been added.
-    throw new UnimplementedError('entries');
-  }
+  Iterable<MapEntry<K, V>> get entries => _map.entries;
 
   @override
   void forEach(void f(K key, V value)) => _map.forEach(f);
@@ -54,20 +44,7 @@ class CopyOnWriteMap<K, V> implements Map<K, V> {
   int get length => _map.length;
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  Map<K2, V2> map<K2, V2>(Object transform(K key, V value)) {
-    // Change Object to MapEntry<K2, V2> when
-    // the MapEntry class has been added.
-    throw new UnimplementedError('map');
-  }
-
-  @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  Map<K2, V2> retype<K2, V2>() {
-    throw new UnimplementedError('retype');
-  }
+  Map<K2, V2> map<K2, V2>(MapEntry<K2, V2> f(K key, V value)) => _map.map(f);
 
   @override
   Iterable<V> get values => _map.values;
@@ -87,12 +64,9 @@ class CopyOnWriteMap<K, V> implements Map<K, V> {
   }
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  void addEntries(Iterable<Object> entries) {
-    // Change Iterable<Object> to Iterable<MapEntry<K, V>> when
-    // the MapEntry class has been added.
-    throw new UnimplementedError('addEntries');
+  void addEntries(Iterable<MapEntry<K, V>> entries) {
+    _maybeCopyBeforeWrite();
+    _map.addEntries(entries);
   }
 
   @override
@@ -114,27 +88,24 @@ class CopyOnWriteMap<K, V> implements Map<K, V> {
   }
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
   void removeWhere(bool test(K key, V value)) {
-    throw new UnimplementedError('removeWhere');
+    _maybeCopyBeforeWrite();
+    _map.removeWhere(test);
   }
 
   @override
   String toString() => _map.toString();
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
   V update(K key, V update(V value), {V ifAbsent()}) {
-    throw new UnimplementedError('update');
+    _maybeCopyBeforeWrite();
+    return _map.update(key, update, ifAbsent: ifAbsent);
   }
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
   void updateAll(V update(K key, V value)) {
-    throw new UnimplementedError('updateAll');
+    _maybeCopyBeforeWrite();
+    _map.updateAll(update);
   }
 
   // Internal.