+++ lib/src/list/built_list.dart
@@ -95,6 +95,10 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   /// As [List.elementAt].
   E operator [](int index) => _list[index];
 
+  /// As [List.+].
+  BuiltList<E> operator +(BuiltList<E> other) =>
+      new _BuiltList<E>.withSafeList(_list + other._list);
+
   /// As [List.length].
   @override
   int get length => _list.length;
@@ -107,6 +111,14 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
 
   /// As [List.lastIndexOf].
   int lastIndexOf(E element, [int start]) => _list.lastIndexOf(element, start);
+
+  /// As [List.indexWhere].
+  int indexWhere(bool test(E element), [int start = 0]) =>
+      _list.indexWhere(test, start);
+
+  /// As [List.lastIndexWhere].
+  int lastIndexWhere(bool test(E element), [int start]) =>
+      _list.lastIndexWhere(test, start);
 
   /// As [List.sublist] but returns a `BuiltList<E>`.
   BuiltList<E> sublist(int start, [int end]) =>
@@ -130,11 +142,7 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
   Iterable<E> where(bool test(E element)) => _list.where(test);
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  Iterable<T> whereType<T>() {
-    throw new UnimplementedError('whereType');
-  }
+  Iterable<T> whereType<T>() => _list.whereType<T>();
 
   @override
   Iterable<T> expand<T>(Iterable<T> f(E e)) => _list.expand(f);
@@ -215,27 +223,14 @@ abstract class BuiltList<E> implements Iterable<E>, BuiltIterable<E> {
       _list.lastWhere(test, orElse: orElse);
 
   @override
-  E singleWhere(bool test(E element), {E orElse()}) {
-    if (orElse != null) throw new UnimplementedError('singleWhere:orElse');
-    return _list.singleWhere(test);
-  }
+  E singleWhere(bool test(E element), {E orElse()}) =>
+      _list.singleWhere(test, orElse: orElse);
 
   @override
   E elementAt(int index) => _list.elementAt(index);
 
   @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  BuiltList<T> cast<T>() {
-    throw new UnimplementedError('cast');
-  }
-
-  @override
-  // TODO: Dart 2.0 requires this method to be implemented.
-  // ignore: override_on_non_overriding_method
-  BuiltList<T> retype<T>() {
-    throw new UnimplementedError('retype');
-  }
+  Iterable<T> cast<T>() => Iterable.castFrom<E, T>(_list);
 
   // Internal.